requirement.txt
fastapi
uvicorn
pillow
python-dotenv
google-generativeai
